+++ spec/console_kit/tenant_configurator_spec.rb
@@ -15,11 +15,6 @@ RSpec.describe ConsoleKit::TenantConfigurator do
   before do
     # Stub configuration accessor
     allow(ConsoleKit.configuration).to receive_messages(tenants: tenants, context_class: context_class)
-
-    stub_const('ApplicationRecord', Class.new { def self.establish_connection(_arg = nil); end })
-    stub_const('Mongoid', Class.new { def self.override_database(_arg); end })
-    stub_const('Redis', Class.new { def self.current; end })
-    stub_const('Elasticsearch', Module.new)
   end
 
   shared_examples 'prints configuration error' do |expected_message:|